chore: add lint step with eslint (#118)
## Description

This PR adds linting with eslint, which allows small errors in code to be picked up quickly, such as dead code or bad references.

This linting is enabled by adding eslint to this package, and adding a npm script: `npm run lint`. It is also enforced by Travis CI, so any PRs will not be able to break the lint. Furthermore, it is integrated into any editors that developers/contributors will use if they have a compatible extension.

## Steps

Check that CI still passes.