build: add regex to release plugin
This commit changes the semantic-release replace plugin replace text
from a string to a regex. This is more reliable when trying to find the
string in the code base to replace and inline with what's reccomended by
the plugin authors. It also fixes the files directory, which was missing
the `js` sub directory.