build: enable semantic release in all branches
This commit changes the semantic-release replace plugin replace text
from a string to a regex. This is more reliable when trying to find the
string in the code base to replace and inline with what's reccomended by
the plugin authors. It also fixes the files directory, which was missing
the `js` sub directory.

The commit also changes the travil deploy script to target all branches.